+++ src/collective/nitf/static/nitf.js
@@ -4,13 +4,13 @@
         // If this is a NITF content with an image
         var nitf_view = $('body.portaltype-collective-nitf-content.template-view').length > 0;
         var nitf_with_image = nitf_view && $('.newsImageContainer').length > 0;
-        var nitf_galleria = $('body.portaltype-collective-nitf-content.template-nitf_galleria').length > 0;
+        var nitf_galleria = $('body.portaltype-collective-nitf-content.template-galleria').length > 0;
         if (nitf_with_image) {
             // Turn image link into something special
             var link = $(".parent-nitf-image");
             // For mobile devices, link will open galleria
             if (/Android|webOS|iPhone|iPad|iPod|BlackBerry/i.test(navigator.userAgent)) {
-                link.attr('href', link.attr('href').replace('@@nitf_galleria', '@@nitf_galleria?ajax_include_head=1&amp;ajax_load=1'));
+                link.attr('href', link.attr('href').replace('@@galleria', '@@galleria?ajax_include_head=1&amp;ajax_load=1'));
             } else {
                 // For other devices open a galleria in an overlay
                 link.prepOverlay({