+++ src/collective/nitf/tests/test_views.py
@@ -152,12 +152,12 @@ class NewsMLViewTestCase(BaseViewTestCase):
 
 class GalleriaViewTestCase(BaseViewTestCase):
 
-    def test_nitf_galleria_view_is_registered(self):
+    def test_galleria_view_is_registered(self):
         registered = [v.name for v in registration.getViews(INITFLayer)]
-        self.assertIn('nitf_galleria', registered)
+        self.assertIn('galleria', registered)
 
         # raises InvalidParameterError if the view is not registered
-        api.content.get_view(u'nitf_galleria', self.n1, self.request)
+        api.content.get_view(u'galleria', self.n1, self.request)
 
 
 class MediaViewTestCase(BaseViewTestCase):
@@ -218,5 +218,5 @@ class RegisteredViewsTestCase(BaseViewTestCase):
     def test_registered_views(self):
         registered = [v.name for v in registration.getViews(INITFLayer)]
         for view in ('edit', 'nitf', u'l10n.datepicker', 'characters-count.js',
-                     'newsml', 'media', 'nitf_galleria', 'view'):
+                     'newsml', 'media', 'galleria', 'view'):
             self.assertTrue(view in registered)